Refactor DQN
Refactor the dqn algorithm so that it has the following logic:
 1. Make steps in the environment and store the transitions in a buffer.
 2. After every fixed number of steps update the model by sampling
    random batches from the buffer.

Move away from the notion of trajectories. Chaning the number of
environments while also keeping the trajectory length fixed, results in
the agent performing a different number of total steps in the
environment. With off-policy algorithms we want to explicitly control
the number of steps.